Implemented CheckOutButtons containing check-out buttons, the time, and order total.
