+++ components/CheckOutButtons.tsx
@@ -1,0 +1,67 @@
+import { View, Text, StyleSheet } from "react-native"
+import { useState, useEffect } from "react"
+
+import CheckOutButton from "./CheckOutButton"
+
+function CheckOutButtons() {
+  
+  // Display and update the time every second
+  const [date, setDate] = useState(new Date())
+  useEffect(() => {
+    var timer = setInterval(() => setDate(new Date()), 1000)
+    return function cleanup() {
+      clearInterval(timer)
+    }
+  })
+
+  return (
+    <View style={styles.checkOutContainer}>
+      <View style={styles.orderTotalRow}>
+        <Text style={styles.dateTimeText}>
+          {date.toLocaleString("en-CA", {
+            timeZone: "EDT",
+            timeStyle: "short",
+            dateStyle: "medium",
+          })}
+        </Text>
+        <Text style={styles.orderTotalText}>39.90</Text>
+      </View>
+      <View style={styles.checkOutButtonsRow}>
+        <CheckOutButton buttonName="Cash" />
+        <CheckOutButton buttonName="Card" />
+      </View>
+      <View style={styles.checkOutButtonsRow}>
+        <CheckOutButton buttonName="39.90" />
+        <CheckOutButton buttonName="40" />
+        <CheckOutButton buttonName="50" />
+        <CheckOutButton buttonName="60" />
+      </View>
+    </View>
+  )
+}
+
+export default CheckOutButtons
+
+const styles = StyleSheet.create({
+  checkOutContainer: {
+    backgroundColor: "#2F3135",
+    paddingBottom: 5,
+  },
+  checkOutButtonsRow: {
+    flexDirection: "row",
+    paddingHorizontal: 2,
+  },
+  orderTotalRow: {
+    flexDirection: "row",
+    padding: 20,
+    justifyContent: "space-between",
+  },
+  dateTimeText: {
+    color: "white",
+  },
+  orderTotalText: {
+    color: "white",
+    fontSize: 20,
+    fontWeight: "bold",
+  },
+})
